move config from types.
