+++ server/internal/config/desktop.go
