+++ server/internal/config/websocket.go
