+++ server/internal/types/desktop.go
@@ -1,5 +1,17 @@
 package types
 
+type ScreenSize struct {
+	Width  int   `json:"width"`
+	Height int   `json:"height"`
+	Rate   int16 `json:"rate"`
+}
+
+type ScreenConfiguration struct {
+	Width  int           `json:"width"`
+	Height int           `json:"height"`
+	Rates  map[int]int16 `json:"rates"`
+}
+
 type DesktopManager interface {
 	Start()
 	Shutdown() error
