--- server/internal/types/xorg.go
@@ -1,13 +1,0 @@
-package types
-
-type ScreenSize struct {
-	Width  int   `json:"width"`
-	Height int   `json:"height"`
-	Rate   int16 `json:"rate"`
-}
-
-type ScreenConfiguration struct {
-	Width  int           `json:"width"`
-	Height int           `json:"height"`
-	Rates  map[int]int16 `json:"rates"`
-}
