+++ server/internal/webrtc/webrtc.go
@@ -14,8 +14,8 @@ import (
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
+	"m1k1o/neko/internal/config"
 	"m1k1o/neko/internal/types"
-	"m1k1o/neko/internal/types/config"
 )
 
 func New(sessions types.SessionManager, capture types.CaptureManager, desktop types.DesktopManager, config *config.WebRTC) *WebRTCManager {
